Make initial database connection for location page

+++ mobile_application/app/src/main/java/com/example/aerosense_app/LocationData.kt
@@ -1,6 +1,6 @@
 package com.example.aerosense_app
 
 data class LocationData (
-    val longitude: Float,
-    val latitude: Float,
+    val longitude: Double,
+    val latitude: Double,
 )
